Correct a bug in the description of the WORK parameter in (c,z)la_syrpvgrw.f WORK is real Bug report sent from Elena Ivanova (Oracle) on Feb 14th 2014

+++ SRC/zla_syrpvgrw.f
@@ -104,7 +104,7 @@
 *>
 *> \param[in] WORK
 *> \verbatim
-*>          WORK is COMPLEX*16 array, dimension (2*N)
+*>          WORK is DOUBLE PRECISION array, dimension (2*N)
 *> \endverbatim
 *
 *  Authors:
